requirements.txt
APScheduler~=3.10.1
fastapi~=0.94.1
jose~=1.0.0
loguru~=0.6.0
aerich~=0.7.0
websockets~=12.0
tortoise-orm~=0.20.0
uvicorn~=0.21.0
passlib[bcrypt]
pydantic[dotenv]
python-jose[cryptography]
redis~=5.0.1
python-dotenv~=1.0.0
pymongo~=4.6.1
dnspython~=2.5.0
starlette~=0.26.1
uvicorn~=0.21.0
unicorn~=2.0.1
urllib3~=2.1.0
anyio~=4.2.0
aiofiles~=23.2.0
bcrypt~=4.1.2
chardet~=3.0.4
PyJWT~=2.8.0
aiomysql~=0.0.20
aiomultiprocess~=0.9.0
asyncio~=3.4.3